requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
django-filter>=23.0
django-redis>=5.4
redis>=5.0
psycopg2-binary>=2.9
whitenoise>=6.6
python-dotenv>=1.0
gunicorn>=21.0
drf-spectacular>=0.27
